requirements.txt
Flask~=1.1
requests~=2.22
pynacl~=1.4.0
robohash~=1.1
pagan~=0.4.3
arrow~=0.17.0
peewee==3.14.0
PyYAML==5.3.1
python-telegram-bot==13.1
arrow~=0.17.0
numpy~=1.19.4